requirements.txt
duckduckgo-search==0.9
python-telegram-bot==20.7
python-dotenv==1.0.0
